requirements.txt
litellm>=1.42.0
pydantic>=2.6.0
networkx>=3.2.1
orjson>=3.10.0
jsonschema>=4.21.0
PyYAML>=6.0.2
